feat(gateway): move initGateway in separate file
