+++ lib/gateway.js
@@ -2,6 +2,9 @@
 
 const buildGateway = require('./gateway/build-gateway')
 const { MER_ERR_INVALID_OPTS } = require('./errors')
+const { assignApplicationLifecycleHooksToContext } = require('./hooks')
+const { kHooks } = require('./symbols')
+const { onGatewayReplaceSchemaHandler } = require('./handlers')
 
 function validateGateway (schema, gateway, opts) {
   if (schema || opts.resolvers || opts.loaders) {
@@ -28,7 +31,90 @@ function validateGateway (schema, gateway, opts) {
   }
 }
 
+async function initGateway (gatewayOpts, fastifyGraphQl, app) {
+  let gatewayRetryIntervalTimer = null
+  const retryServicesCount = gatewayOpts && gatewayOpts.retryServicesCount ? gatewayOpts.retryServicesCount : 10
+
+  const retryInterval = gatewayOpts.retryServicesInterval || 3000
+  const gateway = await buildGateway(gatewayOpts, app)
+
+  const serviceMap = Object.values(gateway.serviceMap)
+  const failedMandatoryServices = serviceMap.filter(service => !!service.error && service.mandatory)
+
+  if (failedMandatoryServices.length) {
+    gatewayRetryIntervalTimer = retryServices(retryInterval)
+    gatewayRetryIntervalTimer.unref()
+  }
+
+  let gatewayInterval = null
+
+  if (gatewayOpts.pollingInterval !== undefined) {
+    if (typeof gatewayOpts.pollingInterval === 'number') {
+      gatewayInterval = setInterval(async () => {
+        try {
+          const context = assignApplicationLifecycleHooksToContext({}, fastifyGraphQl[kHooks])
+          const schema = await gateway.refresh()
+          if (schema !== null) {
+            // Trigger onGatewayReplaceSchema hook
+            if (context.onGatewayReplaceSchema !== null) {
+              await onGatewayReplaceSchemaHandler(context, { instance: app, schema })
+            }
+            fastifyGraphQl.replaceSchema(schema)
+          }
+        } catch (error) {
+          app.log.error(error)
+        }
+      }, gatewayOpts.pollingInterval)
+    } else {
+      app.log.warn(`Expected a number for 'gateway.pollingInterval', received: ${typeof gatewayOpts.pollingInterval}`)
+    }
+  }
+
+  app.onClose((fastify, next) => {
+    gateway.close()
+    if (gatewayInterval !== null) {
+      clearInterval(gatewayInterval)
+    }
+    if (gatewayRetryIntervalTimer !== null) {
+      clearInterval(gatewayRetryIntervalTimer)
+    }
+    setImmediate(next)
+  })
+
+  fastifyGraphQl.gateway = gateway
+  return gateway
+
+  function retryServices (interval) {
+    let retryCount = 0
+    let isRetry = true
+
+    return setInterval(async () => {
+      try {
+        if (retryCount === retryServicesCount) {
+          clearInterval(gatewayRetryIntervalTimer)
+          isRetry = false
+        }
+        retryCount++
+
+        const context = assignApplicationLifecycleHooksToContext({}, fastifyGraphQl[kHooks])
+        const schema = await gateway.refresh(isRetry)
+        /* istanbul ignore next */
+        if (schema !== null) {
+          clearInterval(gatewayRetryIntervalTimer)
+          // Trigger onGatewayReplaceSchema hook
+          if (context.onGatewayReplaceSchema !== null) {
+            await onGatewayReplaceSchemaHandler(context, { instance: app, schema })
+          }
+          fastifyGraphQl.replaceSchema(schema)
+        }
+      } catch (error) {
+        app.log.error(error)
+      }
+    }, interval)
+  }
+}
+
 module.exports = {
-  buildGateway,
+  initGateway,
   validateGateway
 }
